+++ octilab/envs/mdp/actions/__init__.py
@@ -6,6 +6,4 @@
 """Various action terms that can be used in the environment."""
 
 from .actions_cfg import *
-from .binary_joint_actions import *
-from .joint_actions import *
-from .non_holonomic_actions import *
+from .task_space_actions import *