+++ octilab/envs/mdp/actions/actions_cfg.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022-2024, The ORBIT Project Developers.
+# Copyright (c) 2022-2024, The Octi and ORBIT Project Developers.
 # All rights reserved.
 #
 # SPDX-License-Identifier: BSD-3-Clause
@@ -6,177 +6,10 @@
 from __future__ import annotations
 
 from dataclasses import MISSING
-
-from omni.isaac.lab.controllers import DifferentialIKControllerCfg
-from omni.isaac.lab.managers.action_manager import ActionTerm, ActionTermCfg
+from omni.isaac.lab.managers.action_manager import ActionTerm
+from omni.isaac.lab.envs.mdp.actions import DifferentialInverseKinematicsActionCfg
+from octilab.envs.mdp.actions import task_space_actions
 from omni.isaac.lab.utils import configclass
-
-from . import binary_joint_actions, joint_actions, non_holonomic_actions, task_space_actions
-
-##
-# Joint actions.
-##
-
-
-@configclass
-class JointActionCfg(ActionTermCfg):
-    """Configuration for the base joint action term.
-
-    See :class:`JointAction` for more details.
-    """
-
-    joint_names: list[str] = MISSING
-    """List of joint names or regex expressions that the action will be mapped to."""
-    scale: float | dict[str, float] = 1.0
-    """Scale factor for the action (float or dict of regex expressions). Defaults to 1.0."""
-    offset: float | dict[str, float] = 0.0
-    """Offset factor for the action (float or dict of regex expressions). Defaults to 0.0."""
-
-
-@configclass
-class JointPositionActionCfg(JointActionCfg):
-    """Configuration for the joint position action term.
-
-    See :class:`JointPositionAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = joint_actions.JointPositionAction
-
-    use_default_offset: bool = True
-    """Whether to use default joint positions configured in the articulation asset as offset.
-    Defaults to True.
-
-    If True, this flag results in overwriting the values of :attr:`offset` to the default joint positions
-    from the articulation asset.
-    """
-
-
-@configclass
-class RelativeJointPositionActionCfg(JointActionCfg):
-    """Configuration for the relative joint position action term.
-
-    See :class:`RelativeJointPositionAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = joint_actions.RelativeJointPositionAction
-
-    use_zero_offset: bool = True
-    """Whether to ignore the offset defined in articulation asset. Defaults to True.
-
-    If True, this flag results in overwriting the values of :attr:`offset` to zero.
-    """
-
-
-@configclass
-class ExponentialMovingAverageJointPositionActionCfg(JointPositionActionCfg):
-    """Configuration for the exponential moving average joint position action term.
-
-    See :class:`ExponentialMovingAverageJointPositionAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = joint_actions.ExponentialMovingAverageJointPositionAction
-
-    weight: float | dict[str, float] = 1.0
-    """The weight for the moving average (float or dict of regex expressions). Defaults to 1.0.
-
-    If set to 1.0, the processed action is applied directly without any moving average window.
-    """
-
-
-@configclass
-class JointVelocityActionCfg(JointActionCfg):
-    """Configuration for the joint velocity action term.
-
-    See :class:`JointVelocityAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = joint_actions.JointVelocityAction
-
-    use_default_offset: bool = True
-    """Whether to use default joint velocities configured in the articulation asset as offset.
-    Defaults to True.
-
-    This overrides the settings from :attr:`offset` if set to True.
-    """
-
-
-@configclass
-class JointEffortActionCfg(JointActionCfg):
-    """Configuration for the joint effort action term.
-
-    See :class:`JointEffortAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = joint_actions.JointEffortAction
-
-
-##
-# Gripper actions.
-##
-
-
-@configclass
-class BinaryJointActionCfg(ActionTermCfg):
-    """Configuration for the base binary joint action term.
-
-    See :class:`BinaryJointAction` for more details.
-    """
-
-    joint_names: list[str] = MISSING
-    """List of joint names or regex expressions that the action will be mapped to."""
-    open_command_expr: dict[str, float] = MISSING
-    """The joint command to move to *open* configuration."""
-    close_command_expr: dict[str, float] = MISSING
-    """The joint command to move to *close* configuration."""
-
-
-@configclass
-class BinaryJointPositionActionCfg(BinaryJointActionCfg):
-    """Configuration for the binary joint position action term.
-
-    See :class:`BinaryJointPositionAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = binary_joint_actions.BinaryJointPositionAction
-
-
-@configclass
-class BinaryJointVelocityActionCfg(BinaryJointActionCfg):
-    """Configuration for the binary joint velocity action term.
-
-    See :class:`BinaryJointVelocityAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = binary_joint_actions.BinaryJointVelocityAction
-
-
-##
-# Non-holonomic actions.
-##
-
-
-@configclass
-class NonHolonomicActionCfg(ActionTermCfg):
-    """Configuration for the non-holonomic action term with dummy joints at the base.
-
-    See :class:`NonHolonomicAction` for more details.
-    """
-
-    class_type: type[ActionTerm] = non_holonomic_actions.NonHolonomicAction
-
-    body_name: str = MISSING
-    """Name of the body which has the dummy mechanism connected to."""
-    x_joint_name: str = MISSING
-    """The dummy joint name in the x direction."""
-    y_joint_name: str = MISSING
-    """The dummy joint name in the y direction."""
-    yaw_joint_name: str = MISSING
-    """The dummy joint name in the yaw direction."""
-    scale: tuple[float, float] = (1.0, 1.0)
-    """Scale factor for the action. Defaults to (1.0, 1.0)."""
-    offset: tuple[float, float] = (0.0, 0.0)
-    """Offset factor for the action. Defaults to (0.0, 0.0)."""
-
 
 ##
 # Task-space Actions.
@@ -184,36 +17,16 @@ class NonHolonomicActionCfg(ActionTermCfg):
 
 
 @configclass
-class DifferentialInverseKinematicsActionCfg(ActionTermCfg):
-    """Configuration for inverse differential kinematics action term.
+class MultiConstraintsDifferentialInverseKinematicsActionCfg(DifferentialInverseKinematicsActionCfg):
+    """Configuration for inverse differential kinematics action term with multi constraints.
+    This class amend attr body_name from type:str to type:list[str] reflecting its capability to
+    received the desired positions, poses from multiple target bodies. This will be particularly
+    useful for controlling dextrous hand robot with only positions of multiple key frame positions
+    and poses, and output joint positions that satisfy key frame position/pose constrains
 
     See :class:`DifferentialInverseKinematicsAction` for more details.
     """
 
-    @configclass
-    class OffsetCfg:
-        """The offset pose from parent frame to child frame.
-
-        On many robots, end-effector frames are fictitious frames that do not have a corresponding
-        rigid body. In such cases, it is easier to define this transform w.r.t. their parent rigid body.
-        For instance, for the Franka Emika arm, the end-effector is defined at an offset to the the
-        "panda_hand" frame.
-        """
-
-        pos: tuple[float, float, float] = (0.0, 0.0, 0.0)
-        """Translation w.r.t. the parent frame. Defaults to (0.0, 0.0, 0.0)."""
-        rot: tuple[float, float, float, float] = (1.0, 0.0, 0.0, 0.0)
-        """Quaternion rotation ``(w, x, y, z)`` w.r.t. the parent frame. Defaults to (1.0, 0.0, 0.0, 0.0)."""
-
-    class_type: type[ActionTerm] = task_space_actions.DifferentialInverseKinematicsAction
-
-    joint_names: list[str] = MISSING
-    """List of joint names or regex expressions that the action will be mapped to."""
     body_name: list[str] = MISSING
-    """Name of the body or frame for which IK is performed."""
-    body_offset: OffsetCfg | None = None
-    """Offset of target frame w.r.t. to the body frame. Defaults to None, in which case no offset is applied."""
-    scale: float | tuple[float, ...] = 1.0
-    """Scale factor for the action. Defaults to 1.0."""
-    controller: DifferentialIKControllerCfg = MISSING
-    """The configuration for the differential IK controller."""
+
+    class_type: type[ActionTerm] = task_space_actions.MultiConstraintDifferentialInverseKinematicsAction