+++ docs/CHANGELOG.rst
@@ -1,6 +1,23 @@
 Changelog
 ---------
 
+0.1.2 (2024-06-27)
+~~~~~~~~~~~~~~~~~~
+
+Added
+^^^^^
+
+* Added :class:`octilab.envs.mdp.actions.MultiConstraintsDifferentialInverseKinematicsActionCfg`
+
+
+Changed
+^^^^^^^
+* Removed duplicate functions in :class:`octilab.envs.mdp.actions.actions_cfg` already defined in Isaac lab
+* Removed :file:`octilab.envs.mdp.actions.binary_joint_actions.py` as it completely duplicates Isaac lab implementation
+* Removed :file:`octilab.envs.mdp.actions.joint_actions.py` as it completely duplicates Isaac lab implementation
+* Removed :file:`octilab.envs.mdp.actions.non_holonomic_actions.py` as it completely duplicates Isaac lab implementation
+* Cleaned :class:`octilab.controllers.differential_ik.DifferentialIKController`
+
 0.1.1 (2024-06-26)
 ~~~~~~~~~~~~~~~~~~
 
@@ -8,6 +25,14 @@ Added
 ^^^^^
 
 * Rokoko smart glove device reading
+* separation of :class:`octilab.envs.mdp.actions.MultiConstraintDifferentialInverseKinematicsAction` 
+  from :class:`omni.isaac.lab.envs.mdp.actions.DifferentialInverseKinematicsAction`
+
+* separation of :class:`octilab.envs.mdp.actions.MultiConstraintDifferentialIKController` 
+  from :class:`omni.isaac.lab.envs.mdp.actions.DifferentialIKController`
+
+* separation of :class:`octilab.envs.mdp.actions.MultiConstraintDifferentialIKControllerCfg` 
+  from :class:`omni.isaac.lab.envs.mdp.actions.DifferentialIKControllerCfg`
 
 
 Changed
@@ -22,7 +47,7 @@ Changed
 Updated
 ^^^^^^^
 
-* Updated :func:`octilab.envs.DeformableBasedEnv.__init__` up to date with :class:`omni.isaac.lab.envs.ManagerBasedEnv.__init__`
+* Updated :meth:`octilab.envs.DeformableBasedEnv.__init__` up to date with :meth:`omni.isaac.lab.envs.ManagerBasedEnv.__init__`
 * Updated :class:`octilab.envs.HebiRlEnvCfg` to :class:`octilab.envs.OctiManagerBasedRlCfg`  
 * Updated :class:`octilab.envs.HebiRlEnv` to :class:`octilab.envs.OctiManagerBasedRl`